+++ src/constants.rs
@@ -199,9 +199,6 @@ pub(crate) const EDITED_PREFIX: &str = "✏️";
 /// Period between `sql::housekeeping()` runs.
 pub(crate) const HOUSEKEEPING_PERIOD: i64 = 24 * 60 * 60;
 
-/// Seconds between sending out accumulated reaction updates for broadcast channels from `reactions_need_broadcast` table
-pub(crate) const REACTION_BROADCAST_PERIOD: i64 = 10 * 60;
-
 pub(crate) const BROADCAST_INCOMPATIBILITY_MSG: &str = r#"The up to now "experimental channels feature" is about to become an officially supported one. By that, privacy will be improved, it will become faster, and less traffic will be consumed.
 
 As we do not guarantee feature-stability for such experiments, this means, that you will need to create the channel again.
