+++ src/reaction.rs
@@ -14,6 +14,8 @@
 //! possible to remove the reaction by sending an empty string as a reaction,
 //! even though RFC 9078 requires at least one emoji to be sent.
 
+pub(crate) mod broadcast_reactions;
+
 use std::cmp::Ordering;
 use std::collections::BTreeMap;
 use std::fmt;
@@ -27,7 +29,6 @@ use crate::constants::Chattype;
 use crate::contact::ContactId;
 use crate::context::Context;
 use crate::events::EventType;
-use crate::log::warn;
 use crate::message::{Message, MsgId, rfc724_mid_exists};
 use crate::param::Param;
 
@@ -436,103 +437,6 @@ pub async fn get_msg_reactions(context: &Context, msg_id: MsgId) -> Result<React
         .await?;
     reactions.retain(|_contact, reaction| !reaction.is_empty());
     Ok(Reactions { reactions })
-}
-
-/// Sends out accumulated reactions
-/// for all broadcast channels with reactions in `reactions_need_broadcast`.
-///
-/// For every affected `chat_id`,
-/// a single hidden message is sent to all subscribers containing the full, current reaction state (not a diff)
-/// for every message that received a reaction change since the last broadcast.
-pub(crate) async fn broadcast_reactions_for_all_chats(context: &Context) -> Result<()> {
-    let chat_ids: Vec<ChatId> = context
-        .sql
-        .query_map_collect(
-            "SELECT DISTINCT chat_id FROM reactions_need_broadcast",
-            (),
-            |row| {
-                let chat_id: ChatId = row.get(0)?;
-                Ok(chat_id)
-            },
-        )
-        .await?;
-
-    for chat_id in chat_ids {
-        if let Err(err) = broadcast_reactions_for_one_chat(context, chat_id).await {
-            warn!(
-                context,
-                "Failed to broadcast reactions for chat {chat_id}: {err:#}."
-            );
-        }
-    }
-    Ok(())
-}
-
-/// Sends out accumulated reactions for a single broadcast channel
-async fn broadcast_reactions_for_one_chat(context: &Context, chat_id: ChatId) -> Result<()> {
-    let msg_ids: Vec<MsgId> = context
-        .sql
-        .query_map_collect(
-            "SELECT DISTINCT msg_id FROM reactions_need_broadcast WHERE chat_id=?",
-            (chat_id,),
-            |row| {
-                let msg_id: MsgId = row.get(0)?;
-                Ok(msg_id)
-            },
-        )
-        .await?;
-
-    context
-        .sql
-        .execute(
-            "DELETE FROM reactions_need_broadcast WHERE chat_id=?",
-            (chat_id,),
-        )
-        .await?;
-
-    let mut messages: Vec<BroadcastReactionsMessage> = Vec::new();
-    for msg_id in &msg_ids {
-        let Some(msg) = Message::load_from_db_optional(context, *msg_id).await? else {
-            continue;
-        };
-        let reactions = get_msg_reactions(context, *msg_id).await?;
-        let entries: Vec<BroadcastReactionsEntry> = reactions
-            .emoji_sorted_by_frequency()
-            .into_iter()
-            .map(|(emoji, count)| BroadcastReactionsEntry { emoji, count })
-            .collect();
-        messages.push(BroadcastReactionsMessage {
-            id: msg.rfc724_mid,
-            reactions: entries, // can be empty if all reactions were removed
-        });
-    }
-    if messages.is_empty() {
-        return Ok(());
-    }
-
-    let payload = BroadcastReactionsPayload { messages };
-    let json = serde_json::to_string(&payload)?;
-    let mut msg = Message::new_text(json); // TOOD: maybe json is better put to the header
-    msg.hidden = true;
-    send_msg(context, chat_id, &mut msg).await?;
-
-    Ok(())
-}
-
-/// Wire format for accumulated reactions
-#[derive(Debug, Serialize, Deserialize)]
-struct BroadcastReactionsPayload {
-    messages: Vec<BroadcastReactionsMessage>,
-}
-#[derive(Debug, Serialize, Deserialize)]
-struct BroadcastReactionsMessage {
-    id: String,
-    reactions: Vec<BroadcastReactionsEntry>,
-}
-#[derive(Debug, Serialize, Deserialize)]
-struct BroadcastReactionsEntry {
-    emoji: String,
-    count: usize,
 }
 
 impl Chat {
